Created a New File for VaccinationRecord
